# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estacionamento API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/parking": {
            "get": {
                "tags": [
                    "Estacionamento"
                ],
                "summary": "Buscar todas as informações",
                "operationId": "0860ca8ede2e9d07b0fee76d4c7478e7",
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ParkingResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Estacionamento"
                ],
                "summary": "Criar novo estacionamento",
                "operationId": "6dd607e2f83930164d6ed33eb3af6875",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParkingDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParkingResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/parking/{id}": {
            "get": {
                "tags": [
                    "Estacionamento"
                ],
                "summary": "Buscar estacionamento pelo id",
                "operationId": "8824adde947cc86432f5f581ec21259d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParkingResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Estacionamento"
                ],
                "summary": "Deletar por ID",
                "operationId": "b0545907c9f2311287f7c69d193416f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Estacionamento"
                ],
                "summary": "Atualizar estacionamento pelo id",
                "operationId": "cec41ce10b0e4f39189a589cd2b39ee9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParkingDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParkingResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ParkingDTO": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Estacionamento A"
                    },
                    "numberOfVacancies": {
                        "type": "integer",
                        "example": 50
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "ParkingResource": {
                "properties": {
                    "errors": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "content": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "nome": {
                                "type": "string"
                            },
                            "numero_de_vagas": {
                                "type": "integer"
                            },
                            "ativo": {
                                "type": "boolean"
                            },
                            "criado": {
                                "type": "string",
                                "format": "date"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Autenticação",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}